Finished project

+++ src/main/java/repository/UserBookRepository.java
@@ -8,5 +8,5 @@ import domain.UserBook;
 public interface UserBookRepository extends CrudRepository<UserBook, Long>{
 	
 	@Query(value = "SELECT count(*) FROM user_book WHERE book_id = :bookId AND is_favorite = 1", nativeQuery = true)
-	public long findBookRating(Long bookId);
+	public int findBookRating(Long bookId);
 }
